+++ docs/README.md
@@ -1,4 +1,4 @@
-# Alvis.Alvis User Guide
+# screwllum.screwllum User Guide
 
 // Update the title above to match the actual product name
