+++ src/main/java/screwllum/tasks/Event.java
@@ -1,4 +1,4 @@
-package alvis.tasks;
+package screwllum.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
