+++ src/main/java/screwllum/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package alvis.tasks;
+package screwllum.tasks;
 
 public class ToDo extends Task {
     public ToDo(String desc) {
